backend/requirements.txt
Flask==2.2.3
werkzeug==2.2.3
gunicorn==23.0.0
psycopg2-binary==2.9.1  # PostgreSQL 연결을 위한 라이브러리
pyjwt==2.3.0
flask-cors
flask-migrate
flask-sqlalchemy
requests
numpy
pandas
matplotlib
schedule

